Implement the insert operator
Placing the cursor anywhere on the document and typing inserts the
characters into the document.

Also fixed the missing scroll bar at the bottom of the text viewer.
It's not that the items weren't the right size to spawn the scrollbar,
The whole text viewing rectangle was extending below the screen. Used
some anchor magic to make it fit correctly.